### 2018-07-26 25a4046
Update paper.md
