+++ paper/paper.md
@@ -31,15 +31,12 @@ GBs are geometrical entities with a large parameter space that has been well for
 ``GB_code`` is mainly designed to be run in terminal as it is documented in detail in the README file of the repository
 but it can also be accessed via the attached Jupyter notebooks. The code consists of two main scripts, ``csl_generator.py`` and ``gb_generator.py``, that should be used in this order to produce the final GB structures. The attached Jupyter notebooks, ``Usage_of_GB_code.ipynb`` and ``Dichromatic_pattern_CSL.ipynb``, can access the two scripts as modules, the former addresses the
 general usage of the code with some extra tips and functions to locate GBs of interest, the latter depicts how CSL properties such
-as the overlapping patterns and displacement shift complete (DSC) vectors can be extracted and visualized. In the notebooks, two examples of the usage of the ``GB_code`` in our previous publications[@Pub1, @Pub2] have been
-shown as well.
+as the overlapping patterns and displacement shift complete (DSC) vectors can be extracted and visualized. In the notebooks, two examples of the usage of the ``GB_code`` in our previous publications[@Pub1, @Pub2] have been shown as well.
 
 ``GB_code``uses the analytical and mathematical formulations of the following works [@Sutton:1996, @Bollmann:1984, @Grimmer]. Some functionality from this code [@Marcin] on CSL has been used in a modified form in the ``GB_code``. To our knowledge, in comparison to other GB generation codes in different scientific groups``GB_code``is faster due its extensive usage of python Numpy library and is more comprehensive. The user need only to select a rotation axis and then will be guided through in a step by step manner how to find and create the GB of interest. The code has been designed to be simple to use and instructive with a special attention to GB plane orientation which is often lacking in other grain bundary creation codes.
-
 
 # Acknowledgements
 
 R. Hadian would like to thank professor Mike Finnis for fruitful discussions. Funding from the European Union’s Horizon 2020 research and innovation programme (Grant agreement No. 639211) is also gratefully acknowledged.
-
 
 # References
