### 2018-07-26 #853 @oekosheri
Thanks, I edited paper.md and merged. does it look alright now?
