### 2018-07-27 fc40cf0
Update paper.bib
