+++ paper/paper.bib
@@ -1,17 +1,22 @@
-@book{Sutton,
-	Author = {Sutton, A.P. and Balluffi, R.W.},
-	Publisher = {Clarendon Press},
-	Title = {Interfaces in Crystalline Materials},
-	Year = {1996}}
-
-@book{Bollmann,
-author = {Bollmann, W},
-year = {1984},
-month = {01},
-pages = {},
-title = {Crystal Lattices, Interfaces, Matrices: An Extension of Crystallography},
-booktitle = {SERBIULA (sistema Librum 2.0)}
+@book{Sutton:1996,
+  title={Interfaces in Crystalline Materials},
+  author={Sutton, A.P. and Balluffi, R.W.},
+  isbn={9780198500612},
+  lccn={94024731},
+  url={https://books.google.de/books?id=DMafQgAACAAJ},
+  year={1996},
+  publisher={Clarendon Press}
 }
+
+@book{Bollmann:1982,
+  title={Crystal Lattices, Interfaces, Matrices: An Extension of Crystallography},
+  author={Bollmann, W.},
+  isbn={9782881050008},
+  url={https://books.google.de/books?id=oBt0QgAACAAJ},
+  year={1982},
+  publisher={W. Bollmann}
+}
+
 
 @article{LAMMPS,
 	Author = {Plimpton, Steve},
