### 2018-07-27 edfcc90
Update paper.md
