### 2018-08-01 971328b
Update paper.bib
