+++ paper/paper.bib
@@ -7,7 +7,6 @@
   year={1996},
   publisher={Clarendon Press}
 }
-
 @book{Bollmann:1982,
   title={Crystal Lattices, Interfaces, Matrices: An Extension of Crystallography},
   author={Bollmann, W.},
@@ -16,8 +15,6 @@
   year={1982},
   publisher={W. Bollmann}
 }
-
-
 @article{LAMMPS,
 	Author = {Plimpton, Steve},
 	Journal = {Journal of Computational Physics},
@@ -41,7 +38,6 @@
  doi =           {10.1103/PhysRevB.54.11169},
  url =           {http://link.aps.org/doi/10.1103/PhysRevB.54.11169},
 }
-
 @article{Grimmer,
    author = {Grimmer, H. and Bollmann, W. and Warrington, D. H.},
    title = {Coincidence-site lattices and complete pattern-shift in cubic crystals},
@@ -51,7 +47,6 @@
    pages = {197-207},
    year = {1974}
 }
-
 @article{Pub1,
    author = {Hadian, R. and Grabowski, B. and Race, C. P. and Neugebauer, J.},
    title = {Atomistic migration mechanisms of atomically flat, stepped, and kinked grain boundaries},
@@ -68,7 +63,6 @@
    number = {4},
    year = {2018}
 }
-
 @misc{GB_code,
   author = {Sherri Hadian},
   title = {GB_code},
@@ -77,7 +71,6 @@
   journal = {GitHub repository},
   howpublished = {\url{https://github.com/oekosheri/GB_code}},
  }
-
 @misc{Marcin,
   author = {Marcin Wojdyr},
   title = {gosam},
