### 2018-08-01 #853 @oekosheri
@arfon I used pandoc to generate the PDF myself and had no issues. Waiting for your help to see what the problem is.
