### 2018-08-01 f1d6998
Fixing paper compilation
