### 2018-08-01 #853 @oekosheri
Thank you, I merged it. Let me see if I can generate the PDF now.
