### 2018-08-09 #853 @oekosheri
@arfon Dear Editor, I was wondering about the time frame of the review process. Could you let me know approximately when my code will be reviewed? Thanks and I appreciate it.
