### 2018-08-09 #853 @labarba
It can take between two weeks and two months, typically, to complete the review and acceptance process. Thank you for your patience with our wholly volunteer-staffed journal.
