### 2018-08-20 #900 @vyasr
@labarba I'm not quite sure how to review aspects of this repository. It's written in Python, but it's not really a Python package, it's a collection of scripts for a particular purpose. There is no Python API, the intent is just to run these scripts on the command line. That being the case, the normal aspects of a Python package (a setup.py for installation, a Python API, etc) don't really apply. I can request tests for the command line API and some versioning on the repository itself, but since the package won't be `setup.py installed` the version would also just be viewed by looking at the repository source.

Should I just treat this as a command line tool, and give advice accordingly? Or should I be providing suggestions on how to expose/improve a Python API and follow more standard Python packaging guidelines? @oekosheri do you want/see a use for making this tool scriptable within Python rather than as a standalone set of scripts?
