### 2018-08-20 #900 @oekosheri
True, I have not provided a setup.py but this is definitely a python package.
Both main scripts are modules (collection of functions and classes) as well as command line scripts because of the main functions in the end that parses input arguments. Why I have chosen this structure: historical reasons + the underlying physics + preference by the scientists who use it.
If you think I should use a setup.py, I can add it of course. If you have suggestions as to how it should be written, let me know please. If you meant something else that I didn't understand correctly please clarify it a bit more to me.
