### 2018-08-20 #900 @vyasr
You understood my comment correctly. The underlying code is cleanly divided into functions and classes, and you are correct that in the strict sense the repository is a Python package since it is importable thanks to the __init__.py file. I was mostly alluding to the fact that it is not immediately installable, and that the intended API appears to be entirely command line based. You've put together a nice implementation of a cool scientific application, I'm just trying to get a sense for how to make it most useful.

My recommendations will depend a little bit on how you anticipate people making use of the code. If your intended usage is purely from the command line, then I think my comments regarding how to appropriately evaluate this apply because there really isn't much of an API other than simply executing the scripts as you demonstrate in your README. On the other hand, if you want people to be able to interface directly with the Python functions that you provide (in my opinion, you should allow that since it gives users much more flexibility and they might find use-cases you didn't anticipate), then there are a few things you would want to do. Your repository should be restructured  in the form of a standard Python package (see [here](https://docs.python-guide.org/writing/structure/) for example) and you would use a setup.py to make it installable and accessible. If you want to expose a command line interface, the way to do it in that context would be using setuptools entry_points (if you're not familiar, you can see the [official docs](https://setuptools.readthedocs.io/en/latest/setuptools.html#eggsecutable-scripts) as well as a few [basic](https://amir.rachum.com/blog/2017/07/28/python-entry-points/) [tutorials](https://www.eflglobal.com/setuptools-entry-points/)). Even if you don't provide a Python API, making your scripts available on the command line in this fashion would make it easier for people to incorporate the commands into their own (Python, bash, etc) scripts.

Let me know your thoughts on these ideas and I can tailor my feedback accordingly.
