### 2018-08-21 #900 @oekosheri
Thank you for your help.
To answer your question:
Currently, the more classic scientists use it via the Linux terminal and the younger scientists use the Jupyter notebooks. They can import different modules for the functionality they need and there are also additional functions in the notebooks.  As far as the further development goes, I see only myself as the only developer of this code.

1- you say: _"Your repository should be restructured in the form of a standard Python package "_, I do not quite understand how. Do you mean to separate the script part (main functions) of the codes from the modules (functions/class)?
If your answer is yes, I must say I would prefer to keep them as is because of historical preference.
If your answer is different, please let me know.

2- I think it would probably make more sense to me to use the "entry point" solution for the setup.py and keep the scripts as they are.
