### 2018-08-21 #900 @vyasr
As you pointed out, your code already is a package. The standard structure for Python packages, however, is to have all the code in a subdirectory of the repository:
README.rst
LICENSE
setup.py
requirements.txt
sample/__init__.py
sample/core.py
sample/helpers.py
docs/conf.py
docs/index.rst
tests/test_basic.py
tests/test_advanced.py
