### 2018-08-22 #900 @oekosheri
Yes, that's right. thanks.
 So I'm trying re-structuring and I have a question. I put the two scripts/modules in gb_code directory. My setup.py file looks like this:

```
 from setuptools import setup

packages = [
    'gb_code',
       ]
setup(
    name='GB_code',
    version='0.1.0',
    author='R.Hadian',
    author_email='shahrzadhadian@gmail.com',
    packages=packages,
    entry_points = {
        'console_scripts': [
            'csl_generator = gb_code.csl_generator:main',
            'gb_generator = gb_code.gb_generator:main',
        ],
    },
    url='https://github.com/oekosheri/GB_code',
    license='LICENSE',
    description='A GB generation code',
    long_description=open('README.md').read(),
)
```
It copies the executable scripts in python bin, so very good but it does not copy them in site-packages.
if instead of _setuptools_ I use _distutils.core_ the modules will be installed in site-packages but the entry points are not recognised. Do you know what may be the issue?

Thanks in advance.
