### 2018-08-22 #900 @oekosheri
A follow up on my previous post:
I tried``` pip install . ``` and that did the task. i.e. package was copied to my python site-packages and executables to python bin. I wonder if this is the best practice though. If you know of a better way, let me know.
