### 2018-08-22 #900 @vyasr
I don't know exactly what caused the issue that you were having, but in general setup.py is known to have some buggy behavior in odd cases like that, so I don't think that you're doing anything wrong. You could try on a different python version if you have conda/virtualenv handy and see how that behaves, but I wouldn't be too worried.
