### 2018-08-23 7f9275d
re-structuring, changing README accordingly
