### 2018-09-06 #900 @vyasr
Thanks for those changes! Sorry for the delay. Here is my next set of comments.

- Python functions should add some function/class level documentation of the APIs. The functions all have docstrings, which is helpful, but they should also have some description of what the arguments are and what the object types are (floats, lists, some internal class, etc).
- Since you have a setup.py already and your project is pure Python, you can make it even easier to install by uploading your project to PyPI (or even Anaconda cloud). For more information on this, see the [Python Packaging Guide](https://packaging.python.org/), and specifically the guide on [managing and packaging your own project](https://packaging.python.org/tutorials/packaging-projects/).
- Can you add some tests to make sure that the outputs are as expected? I can run your code, but I don't have great topical knowledge of CSL so it would be helpful if you added a tests folder that would run the generator and then check that the output files are what you would expect from theory. For this, you can use the Python [unittest](https://docs.python.org/3/library/unittest.html) framework, or [pytest](https://docs.pytest.org/en/latest/) if you want something more sophisticated.
