### 2018-09-10 #900 @oekosheri
Thank you for your comments.
- I will add more description to the docstrings to describe the parameters a bit more but I hope I don't have to necessarily have full descriptions the same way python does for example. That would take a very long time.
- I  think I can consider this later but for now the installation is rather easy and it seem to work fine.
- The reason I added the Jupyter notebooks to this directory is that they technically show the functionality aside from the command line. So I thought that should serve the purpose of testing as well. If I must add the unittest I will but If the notebooks suffice that would be my preference.
- For a variety of reasons I do not want to accept contribution to the code at this point. I plan to expand this code and when the time comes I will add a contribution section to the README. I hope this is alright.

I submitted the code more than two months ago and have a deadline for presenting it before leaving my institute. Prior to submitting it, I have put quite some time and effort into making the functionality and presentation of it match the needs of current/future users. I really appreciate the fact that I could improve the installation procedure following your comments last time but I also have a time limit now. Aside from completing the docstrings which I will try to finish tomorrow, I would appreciate it if you could test the code now as is.
