### 2018-09-12 f1fbfad
Adding to documentation.
