### 2018-09-12 2f74bf4
Changed the module description.
