+++ gb_code/gb_generator.py
@@ -17,7 +17,7 @@ import sys
 import numpy as np
 from numpy import dot, cross
 from numpy.linalg import det, norm
-import gb_code.csl_generator as cslgen
+import csl_generator as cslgen
 import warnings
 
 
