### 2018-09-12 f9ce97e
pylint changes.
