+++ gb_code/csl_generator.py
@@ -94,11 +94,11 @@ def rot(a, Theta):
     a = a / norm(a)
     ax, ay, az = a
     return np.array([[c + ax * ax * (1 - c), ax * ay * (1 - c) - az * s,
-        ax * az * (1 - c) + ay * s],
-        [ay * ax * (1 - c) + az * s, c + ay * ay * (1 - c),
-        ay * az * (1 - c) - ax * s],
-        [az * ax * (1 - c) - ay * s, az * ay * (1 - c) + ax * s,
-        c + az * az * (1 - c)]])
+                      ax * az * (1 - c) + ay * s],
+                    [ay * ax * (1 - c) + az * s, c + ay * ay * (1 - c),
+                        ay * az * (1 - c) - ax * s],
+                     [az * ax * (1 - c) - ay * s, az * ay * (1 - c) + ax * s,
+                      c + az * az * (1 - c)]])
 
 
 # Helpful Functions:
@@ -199,7 +199,7 @@ def Tilt_Twist_comp(v1, uvw, m, n):
     else:
         twist = 2 * acos(cos(theta / 2) / cos(radians(tilt / 2)))
         print("Tilt component: {0:<6.2f} Twist component: {1:6.2f}"
-            .format(tilt, twist))
+              .format(tilt, twist))
 
 
 def Create_Possible_GB_Plane_List(uvw, m=5, n=1, lim=5):
@@ -232,10 +232,10 @@ def Create_Possible_GB_Plane_List(uvw, m=5, n=1, lim=5):
                    [1, 1, 0],
                    [0, 1, 1],
                    [1, 0, 1],
-                   ],dtype='float')
+                  ], dtype='float')
     # Find GB plane coordinates:
     for i in range(len(indice_0)):
-        if (CommonDivisor(indice_0[i])[1] <= 1):
+        if CommonDivisor(indice_0[i])[1] <= 1:
             V1[i, :] = (indice_0[i, 0] * Min_1[:, 0] +
                         indice_0[i, 1] * Min_1[:, 1] +
                         indice_0[i, 2] * Min_1[:, 2])
@@ -252,12 +252,12 @@ def Create_Possible_GB_Plane_List(uvw, m=5, n=1, lim=5):
         if ang(V1[i], uvw) < tol:
 
             for j in range(len(SymmEquivalent(MP))):
-                if (1 - ang(MeanPlanes[i], SymmEquivalent(MP)[j]) < tol):
+                if 1 - ang(MeanPlanes[i], SymmEquivalent(MP)[j]) < tol:
                     GBtype.append('Symmetric Tilt')
                     break
             else:
                 GBtype.append('Tilt')
-        elif (1 - ang(V1[i], uvw) < tol):
+        elif 1 - ang(V1[i], uvw) < tol:
             GBtype.append('Twist')
         else:
             GBtype.append('Mixed')
@@ -282,7 +282,7 @@ def Create_minimal_cell_Method_1(sigma, uvw, R):
     # remove 0 vectors and uvw from the list
     indice_0 = indice[np.where(np.sum(abs(indice), axis=1) != 0)]
     condition1 = ((abs(dot(indice_0, uvw) / norm(indice_0, axis=1) /
-                  norm(uvw))).round(7))
+                       norm(uvw))).round(7))
     indice_0 = indice_0[np.where(condition1 != 1)]
 
     if MiniCell_search(indice_0, MiniCell_1, R, sigma):
@@ -308,7 +308,7 @@ def MiniCell_search(indices, MiniCell_1, R, sigma):
     Found = False
     count = 0
     while (not Found) and count < len(TestVecs) - 1:
[…]
