### 2018-09-12 bdde24f
adding more to documentation.
