### 2018-09-12 #900 @oekosheri
@labarba @vyasr @trallard I have updated the repository after adding more to the doctrings of the csl_generator script. I have checked the requirement for the documentation in JOSS and think my code satisfies it for the most part. If there is another necessary step that I'm missing please let me know.
Aside from this, I would like to know the timeline for the review process as I must present the code before my departure.
