### 2018-09-12 2a06471
minor additions
