### 2018-09-16 #900 @vyasr
That's fair enough. I can review as is.

The only remaining requirement I have is that you update the setup.py script to have install_requires = ['numpy']. Currently if you try to install this into a clean environment it will fail to run. The requirements.txt file is helpful, but users using pip/setup.py based installation should get the dependencies automatically. pandas and matplotlib are not necessary, since they're just for the notebooks, not the package itself (as far as I can tell).

Regarding not accepting contributions, as far as I am concerned that is fine. @labarba may have more to say on that front, but I will check that box for now and can revise if needed.

Once the setup change is made I can approve.
