### 2018-09-17 #900 @oekosheri
@vyasr thanks! I updated setup.py file with the required numpy version.
@labarba please let me know whether I should add anything to README regarding the contributions. As stated before, my preference is that for now I will not accept contributions. Later on I can change that.
