### 2018-09-17 oekosheri/GB_code#5 @trallard
[JOSS review] -  documentation enhancements
Since you are using the README as the documentation and per the following response on the JOSS review:
>The reason I added the Jupyter notebooks to this directory is that they technically show the functionality aside from the command line. So I thought that should serve the purpose of testing as well. If I must add the unittest I will but If the notebooks suffice that would be my preference.

I would encourage you to add a link to the Notebook on the README, many people use these as documentation/ways to get started and I think these notebooks do provide a good overview of the package
