### 2018-09-17 #900 @trallard
- [documentation](https://github.com/oekosheri/GB_code/issues/5):  as @vyasr  noticed before your README is your main documentation, but you also have a couple of notebooks in the `test` directory it would be nice to point to their existence
