### 2018-09-17 9fee5fe
added python version to setup.py
