### 2018-09-17 #900 @oekosheri
Thank you @trallard
I specified the version of python in setup.py (additionally in README).
I detailed the status on contribution in README.
Regarding the Jupyter notebooks: I have mentioned them in the README and have linked to them.
I can maybe put them outside Test and on the root directory if they are not visible. What is your suggestion?
