### 2018-09-17 #900 @trallard
Thanks a lot for the quick actions @oekosheri.

I closed the notebooks issue as I realized that I did not notice the link to the notebooks.
But for now I'd leave them where they are
