### 2018-09-17 #900 @labarba
@oekosheri : The note on the README is fine. We discussed among JOSS editors, and the minimum requirement is to have an open issue tracker, and to include mention in the README of the contributions policy, but accepting third-party contributions is not a requirement.

@trallard, @vyasr : Are you both comfortable with recommending acceptance, at this point? I need a clear statement of agreement from you to move forward.
