### 2018-09-18 #900 @oekosheri
@trallard I just added two lines to README from the paper as the statement of need.
@labarba Thanks for the explanation. Please let me know of the next steps I need to take.
@vyasr thank you for your help!
