### 2018-09-18 #900 @labarba
👋 @trallard Awaiting your confirmation that you recommend acceptance, at this point.
