### 2018-09-19 #900 @labarba
@oekosheri The paper currently has one big **Summary** section that includes background, software description, and some indication of the intended audience. This content would benefit from some re-organization to bring out the research application of the software into a "Statement of Need." We request that in the [Author's Guide](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain). The guide doesn't explicitly require this section to be separated explicitly under its own heading, but I really prefer it that way. Many JOSS authors already do this.

So, could you add "Statement of Need" as a heading, and write a paragraph that describes to a non-specialist the research application of this software? Examples and references are useful, but keep it brief.

Note that there are some spaces missing around the software name in the last paragraph.
