### 2018-09-20 bc7e43a
Update paper.md
