### 2018-09-20 #900 @oekosheri
@labarba Thanks for the tips. So I reorganised it as you suggested and made some additional modifications. Please let me know if it looks/reads better now.
