### 2018-09-20 d8a4a41
Update paper.md
