### 2018-09-20 #900 @labarba
* Small edit:  capitalize "Statement" on the heading.

* "in comparison to other GB generation codes in different scientific groups `GB_code` is faster due its extensive usage of python and Numpy library and is more comprehensive" >> several small grammatical issues with this clause, but my main question is whether you can say that `GB_code` "is faster" if you are not providing evidence via a performance comparison. Do you have this evidence somewhere?

Note that Python should be capitalized and NumPy also uses a capital P.
