### 2018-09-20 c8688a4
Update paper.md
