### 2018-09-20 #900 @oekosheri
I made some changes to the paragraph.
The point about being faster: personal usage of other codes (I mean the two that I know, there are not many published general GB generation codes so far as I know) have led me to write it. It becomes increasing more important when creating larger structures.

I added the "To our knowledge" to reduce the emphasis. I can remove it altogether if you think it does not fit.
