### 2018-09-20 #900 @oekosheri
I don't have numbers of course and it is not a crucial point for the paper so I rephrased it according to your suggestion. Please let me know if it is better now. Thank you.
