### 2018-09-20 #900 @labarba
**Editorial fixes**:

`GB_code` is mainly designed to be a command line tool as it is documented >> `GB_code is designed to be a command-line tool, as documented… (delete superfluous adverb, hyphen compound adjective, add comma, delete extra words)

README file of the repository but the modules >> add comma before “but”

Inputs the two scripts as modules, the former >> period instead of comma

Of the following works >> edit the citation to remove the parenthesis around the list of refs.

Some functionality from this code (Wojdyr, 2013) >> makes no sense to say “this code” followed by a reference in parenthesis; please use the code name, or write “from the code by Wonder (2013)”

`GB_code` is an interactive >> interactive

The target audience are  >> The target audience is

Extensive use of NumPy library >> the NumPy library

GB_code >> `GB_code` (consistent formatting)

The user will be guided through in a step by step manner how to find and create the GB of interest. >> The user will be guided in a step-by-step manner on how to find and create… (note the hyphens)

GB plane orientation which is often lacking >> comma before “which”

---

Please add missing DOIs to journal articles in references.
