### 2018-09-21 e7f001f
Update paper.bib
