+++ paper/paper.bib
@@ -22,6 +22,7 @@
 	Pages = {1-19},
 	Title = {Fast Parallel Algorithms for Short-Range Molecular Dynamics},
 	Volume = {117},
+	doi =    {10.1006/jcph.1995.1039},
 	Year = {1995}}
 	
 @article{VASP,
@@ -45,14 +46,16 @@
    volume = {30},
    number = {2},
    pages = {197-207},
-   year = {1974}
-}
+   doi   = {https://doi.org/10.1107/S056773947400043X}
+   year = {1974}}
+   
 @article{Pub1,
    author = {Hadian, R. and Grabowski, B. and Race, C. P. and Neugebauer, J.},
    title = {Atomistic migration mechanisms of atomically flat, stepped, and kinked grain boundaries},
    journal = {Physical Review B},
    volume = {94},
    number = {16},
+   doi    = {https://doi.org/10.1103/PhysRevB.94.165413},
    year = {2016}
 }
 @article{Pub2,
@@ -61,6 +64,7 @@
    journal = {Physical Review Materials},
    volume = {2},
    number = {4},
+   doi    = {https://doi.org/10.1103/PhysRevMaterials.2.043601},
    year = {2018}
 }
 @misc{GB_code,
