### 2018-09-21 #900 @oekosheri
Thanks for the edits.
I corrected them all except:

- Of the following works >> edit the citation to remove the parenthesis around the list of refs.

by parenthesis do you mean the brackets? I have them around all other citations too. Should I remove them all?
