### 2018-09-21 #900 @labarba
When the citation is part of the sentence, then use: Author (year). If the citation is not part of the sentence, then it should appear in parenthesis.
E.g.
1. "We take inspiration from Smith et al. (2006) to ... "
2. "The code implements Theory 1 (Jones 2012) and Theory 2 (Roberts 2014) to obtain ... "

For the syntax to obtain brackets in the right places, see: https://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html
