+++ paper/paper.md
@@ -31,7 +31,7 @@ GBs are geometrical entities with a large parameter space that has been well for
 ``GB_code`` is designed to be a command line tool as it is documented in detail in the README file of the repository,
 but the modules can also be accessed separately for example via the attached Jupyter notebooks. The code consists of two main scripts, ``csl_generator.py`` and ``gb_generator.py``, that should be used in this order to produce the final GB structures. The attached Jupyter notebooks in the Test directory, ``Usage_of_GB_code.ipynb`` and ``Dichromatic_pattern_CSL.ipynb``, input the two scripts as modules. The former addresses the general usage of the code with some extra tips and functions to locate GBs of interest, the latter depicts how CSL properties such as the overlapping patterns and displacement shift complete (DSC) vectors can be extracted and visualized. In the notebooks, two examples of the usage of the ``GB_code`` in our previous publications [@Pub1, @Pub2] have been shown as well.
 
-``GB_code`` uses the analytical and mathematical formulations of the following works [@Sutton:1996, @Bollmann:1982, @Grimmer]. Some functionality from the code by [@Marcin] on CSL has been used in a modified form in the ``GB_code``. 
+``GB_code`` uses the analytical and mathematical formulations of the following works: @Sutton:1996, @Bollmann:1982, @Grimmer. Some functionality from the code by [@Marcin] on CSL has been used in a modified form in the ``GB_code``. 
 
 ### Statement of need:
 ``GB_code`` is an interactive toolbox to learn about grain boundaries and it is versatile for running high-throughput calculations. The target audience is students/scientists of materials science and physics at any level of familiarity with the topic. Extensive use of the NumPy library in ``GB_code`` results in faster execution, especially when computing large structures. The user will be guided through in a step-by-step manner on how to find and create the GB of interest. The code has been designed to be simple to use and instructive with a special attention to GB plane orientation, which is often lacking in other grain boundary creation codes.
