### 2018-09-21 1d9f73d
Update paper.bib
