+++ paper/paper.bib
@@ -68,7 +68,7 @@
    year = {2018}
 }
 @misc{GB_code,
-  author = {Sherri Hadian},
+  author = {Raheleh Hadian},
   title = {GB_code},
   year = {2018},
   publisher = {GitHub},
