### 2018-09-21 104db1a
Update paper.md
