+++ paper/paper.bib
@@ -38,7 +38,7 @@
  year =          {1996},
  doi =           {10.1103/PhysRevB.54.11169},
  url =           {http://link.aps.org/doi/10.1103/PhysRevB.54.11169},
-}
+ }
 @article{Grimmer,
    author = {Grimmer, H. and Bollmann, W. and Warrington, D. H.},
    title = {Coincidence-site lattices and complete pattern-shift in cubic crystals},
@@ -47,7 +47,8 @@
    number = {2},
    pages = {197-207},
    doi   = {https://doi.org/10.1107/S056773947400043X}
-   year = {1974}}
+   year = {1974}
+   }
    
 @article{Pub1,
    author = {Hadian, R. and Grabowski, B. and Race, C. P. and Neugebauer, J.},
@@ -57,7 +58,7 @@
    number = {16},
    doi    = {https://doi.org/10.1103/PhysRevB.94.165413},
    year = {2016}
-}
+   }
 @article{Pub2,
    author = {Hadian, R. and Grabowski, B. and Finnis, M. W. and Neugebauer, J.},
    title = {Migration mechanisms of a faceted grain boundary},
@@ -74,7 +75,7 @@
   publisher = {GitHub},
   journal = {GitHub repository},
   howpublished = {\url{https://github.com/oekosheri/GB_code}},
- }
+  }
 @misc{Marcin,
   author = {Marcin Wojdyr},
   title = {gosam},
@@ -82,4 +83,4 @@
   publisher = {GitHub},
   journal = {GitHub repository},
   howpublished = {\url{https://github.com/wojdyr/gosam/blob/master/csl.py}},
- }
+  }
