### 2018-09-21 61c1248
Update paper.md
