+++ paper/paper.bib
@@ -39,7 +39,7 @@
  doi =           {10.1103/PhysRevB.54.11169},
  url =           {http://link.aps.org/doi/10.1103/PhysRevB.54.11169},
  }
-@article{Grimmer,
+@article{Grimmer:1974,
    author = {Grimmer, H. and Bollmann, W. and Warrington, D. H.},
    title = {Coincidence-site lattices and complete pattern-shift in cubic crystals},
    journal = {Acta Crystallographica Section A},
@@ -76,7 +76,7 @@
   journal = {GitHub repository},
   howpublished = {\url{https://github.com/oekosheri/GB_code}},
   }
-@misc{Marcin,
+@misc{Marcin:2013,
   author = {Marcin Wojdyr},
   title = {gosam},
   year = {2013},
