### 2018-09-21 0b4a814
Update paper.md
