+++ paper/paper.bib
@@ -48,8 +48,7 @@
    pages = {197-207},
    doi   = {https://doi.org/10.1107/S056773947400043X}
    year = {1974}
-   }
-   
+   }   
 @article{Pub1,
    author = {Hadian, R. and Grabowski, B. and Race, C. P. and Neugebauer, J.},
    title = {Atomistic migration mechanisms of atomically flat, stepped, and kinked grain boundaries},
@@ -76,7 +75,7 @@
   journal = {GitHub repository},
   howpublished = {\url{https://github.com/oekosheri/GB_code}},
   }
-@misc{Marcin:2013,
+@misc{Marcin,
   author = {Marcin Wojdyr},
   title = {gosam},
   year = {2013},
