### 2018-09-21 ed3f377
Update paper.bib
