+++ paper/paper.bib
@@ -15,7 +15,7 @@
   year={1982},
   publisher={W. Bollmann}
 }
-@article{LAMMPS,
+@article{LAMMPS:1995,
 	Author = {Plimpton, Steve},
 	Journal = {Journal of Computational Physics},
 	Number = {1},
@@ -25,7 +25,7 @@
 	doi =    {10.1006/jcph.1995.1039},
 	Year = {1995}}
 	
-@article{VASP,
+@article{VASP:1996,
  author =        {Kresse, G. and Furthm\"uller, J.},
  journal =       {Phys. Rev. B},
  month =         {Oct},
@@ -49,7 +49,7 @@
    doi   = {https://doi.org/10.1107/S056773947400043X}
    year = {1974}
    }   
-@article{Pub1,
+@article{Pub1:2016,
    author = {Hadian, R. and Grabowski, B. and Race, C. P. and Neugebauer, J.},
    title = {Atomistic migration mechanisms of atomically flat, stepped, and kinked grain boundaries},
    journal = {Physical Review B},
@@ -58,7 +58,7 @@
    doi    = {https://doi.org/10.1103/PhysRevB.94.165413},
    year = {2016}
    }
-@article{Pub2,
+@article{Pub2:2018,
    author = {Hadian, R. and Grabowski, B. and Finnis, M. W. and Neugebauer, J.},
    title = {Migration mechanisms of a faceted grain boundary},
    journal = {Physical Review Materials},
