+++ paper/paper.bib
@@ -46,7 +46,7 @@
    volume = {30},
    number = {2},
    pages = {197-207},
-   doi   = {https://doi.org/10.1107/S056773947400043X}
+   doi   = {https://doi.org/10.1107/S056773947400043X},
    year = {1974}
    }   
 @article{Pub1:2016,
