### 2018-09-21 6ac6cdb
Update paper.md
