### 2018-09-21 bce6d75
Update paper.bib
