+++ paper/paper.bib
@@ -49,7 +49,7 @@
    doi   = {https://doi.org/10.1107/S056773947400043X},
    year = {1974}
    }   
-@article{Pub1:2016,
+@article{Pub1,
    author = {Hadian, R. and Grabowski, B. and Race, C. P. and Neugebauer, J.},
    title = {Atomistic migration mechanisms of atomically flat, stepped, and kinked grain boundaries},
    journal = {Physical Review B},
@@ -58,7 +58,7 @@
    doi    = {https://doi.org/10.1103/PhysRevB.94.165413},
    year = {2016}
    }
-@article{Pub2:2018,
+@article{Pub2,
    author = {Hadian, R. and Grabowski, B. and Finnis, M. W. and Neugebauer, J.},
    title = {Migration mechanisms of a faceted grain boundary},
    journal = {Physical Review Materials},
@@ -68,7 +68,7 @@
    year = {2018}
 }
 @misc{GB_code,
-  author = {Raheleh Hadian},
+  author = {R. Hadian},
   title = {GB_code},
   year = {2018},
   publisher = {GitHub},
