### 2018-09-21 2cdac20
Update paper.md
