### 2018-09-21 #900 @labarba
Something is wrong with the DOI links. For example, the second reference links to: `https://doi.org/https://doi.org/10.1103/PhysRevMaterials.2.043601`

Please clean up the BibTeX entries so they resolve properly.
