+++ paper/paper.bib
@@ -46,7 +46,7 @@
    volume = {30},
    number = {2},
    pages = {197-207},
-   doi   = {https://doi.org/10.1107/S056773947400043X},
+   doi   = {10.1107/S056773947400043X},
    year = {1974}
    }   
 @article{Pub1,
@@ -55,7 +55,7 @@
    journal = {Physical Review B},
    volume = {94},
    number = {16},
-   doi    = {https://doi.org/10.1103/PhysRevB.94.165413},
+   doi    = {10.1103/PhysRevB.94.165413},
    year = {2016}
    }
 @article{Pub2,
@@ -64,7 +64,7 @@
    journal = {Physical Review Materials},
    volume = {2},
    number = {4},
-   doi    = {https://doi.org/10.1103/PhysRevMaterials.2.043601},
+   doi    = {10.1103/PhysRevMaterials.2.043601},
    year = {2018}
 }
 @misc{GB_code,
