### 2018-09-22 #900 @labarba
Great!

The next step is for you to make an archival copy of the full repository in a service like Zenodo, get a DOI, and tell us the archive DOI here so we can add it to the publication.
