+++ README.md
@@ -183,6 +183,3 @@ Should this change in the future I'll provide detailed information on how to con
 
 # License
 [MIT](./LICENSE).
-
-
-
