+++ README.md
@@ -182,4 +182,6 @@ Should this change in the future I'll provide detailed information on how to con
 
 
 # License
-[MIT](./LICENSE).
+[MIT](./LICENSE).  
+
+
