### 2018-09-23 #900 @oekosheri
Thanks @labarba.
I tried to integrate the repository on Zenodo and it constantly fails. Do you have any instructions on zenodo integration? I can't find any details on JOSS pages.
