### 2018-09-23 #900 @oekosheri
Thanks @labarba, so eventually I managed to integrate the repo with Zenodo and it created a badge with Doi.
[![DOI](https://zenodo.org/badge/140718298.svg)](https://zenodo.org/badge/latestdoi/140718298)
Is it correct?
