### 2018-09-23 #900 @labarba
That DOI resolves correctly, yes. But note that the authors are listed as "oekosheri; Arfon Smith" in the Zenodo entry. Probably you had a pull-request from Arfon that you merged, and Zenodo is taking the authors from the committers, automatically. You need to manually edit the metadata to display the correct author list in Zenodo.
