### 2018-09-23 #900 @labarba
Lastly, I do have a question for you. You've only listed yourself as author in the Zenodo archive. We only see yourself as a committer in the code repository. Yet you have three authors in the JOSS paper. I'm a little worried about courtesy authorship here. The last author is the Director of the Computational Materials Design department at MPIE, and the second author is the group leader of Adaptive Structural Materials. Can you look at the [JOSS authorship guide](https://joss.readthedocs.io/en/latest/search.html?q=authorship&check_keywords=yes&area=default), reflect on this (possibly discuss with your listed co-authors) and make a final determination that the author list is what you all want?

From the JOSS guidelines:

> **Authorship**

>Purely financial (such as being named on an award) and organizational (such as general supervision of a research group) contributions are not considered sufficient for co-authorship of JOSS submissions, but active project direction and other forms of non-code contributions are. The authors themselves assume responsibility for deciding who should be credited with co-authorship, and co-authors must always agree to be listed. In addition, co-authors agree to be accountable for all aspects of the work.
