### 2018-09-23 #900 @oekosheri
I am the sole contributor of the code, which is a collection of work I did in the last couple of years under the active direction of the other two authors. So I can update Zenodo author list and include them as well.
