### 2018-09-23 #900 @labarba
@arfon : This submission is accepted and ready to be published.

@vyasr, @trallard : Thank you for your valuable contribution to JOSS by reviewing this publication!
